@@ -1,6 +1,6 @@
 
 
-from typing import Optional
+
 
 from pydantic import BaseModel
 
@@ -9,4 +9,4 @@ class Transaction(BaseModel):
     user_id: str
     amount: float
     type: str
-    date: Optional[str]
+    date: str | None
